Refactor : 가게 상세보기 컴포넌트 이름 변경 및 즐겨찾기 기능 구현 (#65)
